# Conflicts:
#	arangod/RocksDBEngine/RocksDBCollection.cpp

+++ arangod/Cache/CacheManagerFeature.cpp
@@ -59,7 +59,7 @@ CacheManagerFeature::CacheManagerFeature(
       _cacheSize((TRI_PhysicalMemory >= (static_cast<uint64_t>(4) << 30))
                   ? ((TRI_PhysicalMemory - (static_cast<uint64_t>(2) << 30)) * 0.3)
                   : (256 << 20)),
-      _rebalancingInterval(2 * 1000 * 1000) {
+      _rebalancingInterval(static_cast<uint64_t>(2 * 1000 * 1000)) {
   setOptional(true);
   requiresElevatedPrivileges(false);
   startsAfter("Scheduler");
